+++ app/controllers/users_controller.rb
@@ -16,36 +16,6 @@ class UsersController < ApplicationController
     @users = scope_order(scope).page(params[:page]).per(40)
   end
 
-  # GET /users/export
-  def export
-    @csv_string = CSV.generate do |csv|
-      csv << [
-        "MyApnea ID", "Email", "Full Name", "Number of Surveys Completed",
-        "Last Login", "Login Count", "Forum Posts", "First Forum Post",
-        "Last Forum Post"
-      ]
-      User.current.find_each do |user|
-        first_reply = user.replies.order(:created_at).first
-        last_reply = user.replies.order(:created_at).last
-        row = [
-          user.myapnea_id,
-          user.email,
-          user.full_name,
-          -1,
-          (user.current_sign_in_at ? user.current_sign_in_at.to_date : nil),
-          user.sign_in_count,
-          user.replies.count,
-          (first_reply ? first_reply.created_at.to_date : nil),
-          (last_reply ? last_reply.created_at.to_date : nil)
-        ]
-        csv << row
-      end
-    end
-    filename = "myapnea-users-#{Time.zone.now.strftime("%Y-%m-%d-%Ih%M-%p")}.csv"
-    send_data @csv_string, type: "text/csv; charset=iso-8859-1; header=present",
-                           disposition: "attachment; filename=\"#{filename}\""
-  end
-
   # GET /users/1/edit
   # def edit
   # end