+++ db/migrate/20180716165810_add_indices_to_exports.rb
@@ -1,0 +1,6 @@
+class AddIndicesToExports < ActiveRecord::Migration[5.2]
+  def change
+    add_index :admin_exports, :file_size
+    add_index :admin_exports, :status
+  end
+end